+++ dxr/src/types/error.rs
@@ -2,7 +2,7 @@ use std::borrow::Cow;
 
 use thiserror::Error;
 
-use crate::fault::Fault;
+use crate::types::fault::Fault;
 
 #[derive(Debug, Error)]
 /// error type used for conversion errors between XML-RPC values and Rust values