+++ dxr/src/types/tests/types.rs
@@ -1,7 +1,7 @@
 use chrono::{SubsecRound, Utc};
 use quick_xml::{de::from_str, se::to_string};
 
-use crate::types::*;
+use crate::types::structs::*;
 use crate::XML_RPC_DATE_FORMAT;
 
 #[test]